# api/openapi.yaml
openapi: 3.1.0
info:
  title: Demand Toolkit API
  version: 0.0.5
  description: |
    # Demand Toolkit API

    The Demand Toolkit API provides access to demand forecasting data across multiple
    temporal resolutions and geographic boundaries. This API serves as the data layer
    for the complete Demand Toolkit framework, which includes:

    - **Python Generator**: Creates demand forecasts and scenarios
    - **Node.js API**: Serves forecasting data (this API)
    - **Svelte Explorer**: Interactive web interface for data exploration

    ## Key Features

    - **Multi-format Output**: JSON and GeoJSON support for geographic data
    - **Temporal Flexibility**: Hourly to yearly resolution support
    - **Scenario Management**: Multiple forecasting scenarios with parameter combinations
    - **Geographic Boundaries**: Configurable geographic divisions (counties, municipalities, etc.)
    - **Parameter Validation**: All parameters are validated against available data

    ## Getting Started

    1. Use **GET /config** to understand the dataset configuration
    2. Use **GET /parameters** to discover valid parameter combinations
    3. Use **GET /scenarios** to explore available forecasting scenarios
    4. Use **GET /geographies** to get geographic boundary data

    All legal values for query-field variants are retrieved via **GET /parameters**.

  contact:
    name: Demand Toolkit Team
    url: https://github.com/aidotse/behovskartan
  license:
    name: ISC
...
servers:
...
components:
  parameters:
...
    geoFormat:
      $ref: './parameters.yaml#/components/parameters/geoFormat'
    # Strategy 2 Parameters
...
  schemas:
    GeoFeatureCollection:
      $ref: './schemas/common-schemas.yaml#/components/schemas/GeoFeatureCollection'
...
    Error:
      $ref: './schemas/common-schemas.yaml#/components/schemas/Error'
  responses:
...
    NotFound:
      description: |
        The requested endpoint or resource does not exist on this server.
        Safe to retry after fixing the URL.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
            requestId: "4f1e6a1a-6b7c-4f1a-9c0d-0e7a5e3b4c2d"
...
    ServerError:
      description: |
        The server failed to satisfy the request. The `error` field is a short
        generic message — detailed diagnostics are logged server-side under
        `requestId`. Please include the request ID when reporting bugs.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Query failed"
            requestId: "4f1e6a1a-6b7c-4f1a-9c0d-0e7a5e3b4c2d"
paths:
  /geographies:
    get:
      operationId: getGeographies
      summary: Get geographic boundaries and metadata
      description: |
        Returns a list of geographic boundaries available in the dataset. This endpoint
        supports both JSON format (for metadata) and GeoJSON format (for spatial data
        including geometry definitions).

        **Use Cases:**
        - Get list of available geographic divisions (counties, municipalities, etc.)
        - Retrieve geographic metadata (names, IDs, types)
        - Download GeoJSON for mapping and visualization
        - Understand the geographic scope of the demand forecasting dataset

        The response format is determined by the `format` query parameter.
      parameters:
        - $ref: '#/components/parameters/geoFormat'
      responses:
        '200':
          description: |
            Geographic boundaries and metadata. Response format depends on the `format` parameter:
            - `json` (default): Array of geography metadata objects
            - `geojson`: GeoJSON FeatureCollection with geometry data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Unique identifier for the geographic boundary
                      example: "SE-AB"
                    name:
                      type: string
                      description: Human-readable name of the geographic area
                      example: "Stockholms län"
                    type:
                      type: string
                      description: Type of geographic division
                      example: "county"
                  required: [id, name, type]
              examples:
                counties:
                  summary: Swedish counties
                  value:
                    - id: "SE-AB"
                      name: "Stockholms län"
                      type: "county"
                    - id: "SE-C"
                      name: "Gävleborgs län"
                      type: "county"
            application/geo+json:
              schema:
                $ref: '#/components/schemas/GeoFeatureCollection'
              examples:
                geojson:
                  summary: GeoJSON with county boundaries
                  value:
                    type: "FeatureCollection"
                    features:
                      - type: "Feature"
                        properties:
                          id: "SE-AB"
                          name: "Stockholms län"
                          type: "county"
                        geometry:
                          type: "MultiPolygon"
                          coordinates: [[[18.0, 59.0], [18.1, 59.0], [18.1, 59.1], [18.0, 59.1], [18.0, 59.0]]]
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /config:
    get:
      operationId: getConfig
      summary: Data-generator configuration
      description: |
        Returns the generator configuration that produced the current dataset.
        Baked at API-startup time and cached for 1 hour. Use this to discover
        available segments, geographic granularity, and dataset metadata.
      responses:
        '200':
          description: JSON document
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /scenarios:
    get:
      operationId: getScenarios
      summary: Enumerated list of valid scenario permutations
      description: |
        Returns the list of scenarios this dataset contains. Baked at
        API-startup time and cached for 1 hour.
      responses:
        '200':
          description: JSON array
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /parameters:
    get:
      operationId: getParameters
      summary: Lookup tables for every query-field
      description: |
        Returns the complete lookup tables for `/demand` query parameters:
        base scenarios, segment-level parameter knobs (e.g. `housing_growth`,
        `transport_flex`), valid years, resolutions, aggregations, and
        geographies. Each parameter value carries its own index, label, and
        numeric `value`, so clients can render UIs without re-reading the
        generator source. Baked at API-startup time and cached for 1 hour.
      responses:
        '200':
          description: Complete parameter lookup tables.
          content:
            application/json:
              schema:
                $ref: './schemas/common-schemas.yaml#/components/schemas/ParametersResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /aggregations:
    get:
      operationId: getAggregations
      summary: Available resolution-aggregation combinations
      description: |
        Returns the list of supported (resolution, aggregation) tuples that
        `/demand` will accept. Baked at API-startup time.
      responses:
        '200':
          description: JSON array of aggregation options
          content:
            application/json:
              schema:
                $ref: './schemas/common-schemas.yaml#/components/schemas/AggregationsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...